backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
argon2==0.1.10
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
dnspython==2.6.1
fastapi==0.112.0
h11==0.14.0
idna==3.7
iso8601==2.1.0
motor==3.5.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
pyseto==1.7.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.6
